+++ R/api-rng.R
@@ -303,7 +303,7 @@ nv_rbinom <- jit(
 #' @seealso [nv_sample()] to sample from an arbitrary population.
 #' @examplesIf pjrt::plugins_downloaded()
 #' state <- nv_rng_state(42L)
-#' # Roll 6 dice
+#' # roll 6 dice
 #' result <- nv_sample_int(6, state, 6L)
 #' result$values
 #' @export